+++ __init__.py
@@ -20,10 +20,12 @@ import supybot
 from supybot import world
 
 # Use this for the version of this plugin.
-__version__ = "1.0.0"
+__version__ = "1.1.0"
 
 # XXX Replace this with an appropriate author or supybot.Author instance.
-__author__ = supybot.Author("Barry Suridge", "Alcheri", "barry.suridge@outlook.com")
+__author__ = supybot.Author(
+    "Barry Suridge", "Alcheri", "barry.suridge@outlook.com"
+)
 
 # This is a dictionary mapping supybot.Author instances to lists of
 # contributions.